+++ src/BlockHorizons/BlockSniper/data/TranslationData.php
@@ -45,6 +45,13 @@ class TranslationData{
 		return $languageSelected;
 	}
 
+	public function regenerateLanguageFile() : void{
+		unlink($this->loader->getDataFolder() . "languages/" . $this->loader->config->messageLanguage . ".json");
+		// Always unlink the default translations.
+		unlink($this->loader->getDataFolder() . "languages/en.json");
+		$this->collectTranslations();
+	}
+
 	/**
 	 * @return Loader
 	 */